jsondb/cmd/server/main.go
package main

import (
	"log"
	"os"
	"os/signal"
	"runtime"
	"syscall"

	"jsondb/internal/config"
	"jsondb/internal/server"
)

func main() {
	// Log number of CPUs being used
	log.Printf("Using %d CPU cores", runtime.NumCPU())

	// Load configuration
	cfg, err := config.LoadConfig()
	if err != nil {
		log.Fatalf("Failed to load config: %v", err)
	}

	// Create server instance
	srv, err := server.NewServer(cfg)
	if err != nil {
		log.Fatalf("Failed to create server: %v", err)
	}

	// Start the server
	if err := srv.Start(); err != nil {
		log.Fatalf("Failed to start server: %v", err)
	}

	// Handle graceful shutdown
	sigChan := make(chan os.Signal, 1)
	signal.Notify(sigChan, syscall.SIGINT, syscall.SIGTERM)

	// Wait for shutdown signal
	<-sigChan
	log.Println("Shutting down server...")

	// Stop the server
	if err := srv.Stop(); err != nil {
		log.Printf("Error during shutdown: %v", err)
	}
}
